+++ src/main/kotlin/language/RellLexerAdapter.kt
@@ -1,6 +1,7 @@
 package net.postchain.rellide.jetbrains.language
 
 import net.postchain.rellide.jetbrains.language.parser.RellLexer
+import net.postchain.rellide.jetbrains.language.psi.RellPsiElementTypes
 import org.antlr.intellij.adaptor.lexer.ANTLRLexerAdaptor
 
-class RellLexerAdapter : ANTLRLexerAdaptor(RellLanguage.INSTANCE, RellLexer(null))
+class RellLexerAdapter : ANTLRLexerAdaptor(RellPsiElementTypes.registeredLanguage(), RellLexer(null))